simplify math
this way, there is only one quaternion to combine the new gyro value. This makes the order, in which the axis would be applied otherwise, irrelevant. Since this order would just be a definition anyway, and not built upon a real physical/mathematical reason, there is no incentive to keep a specific order. Arbitrary terms have been omitted