use DShot 600, to reduce errors
if the return packet did not arrive by the time the new packet is being processed (can happen due to jitter in the processing time), then an error is recorded. Faster DShot packets mean the chance is higher that the RPMs have been returned. Down from one error every ~900 frames to 0 errors