Update to Svelte 5 (#53)
This is more like a dump of the merge, rather than actually merging. I spent a day on this update and in case it becomes useful again, it's nice to have it in the history. Also some of the code is a bit cleaner than before (e.g. fewer circular effects that only stop because of some of the code is asynchronous and therefore untracked by svelte 5 (svelte 4 resolved these circular dependencies!). Third reason: This is the base I used to upgrade to Vue as the frontend framework, which is going great. The next merge will be the conversion to vue (which is somehow easier than migrating from svelte 4 to 5, despite the migration script).